+++ pyms/config/__init__.py
@@ -1,0 +1,4 @@
+from pyms.config.conf import get_conf
+from pyms.config.confile import ConfFile
+
+__all__ = ['get_conf', 'ConfFile']
