+++ pyms/utils/__init__.py
@@ -1,0 +1,3 @@
+from pyms.utils.utils import import_from, import_package, check_package_exists
+
+__all__ = ['import_from', 'import_package', 'check_package_exists']
